Removed alt-graphics, new graphics working but lots of testing. Bully not working
