+++ TODO.md
@@ -1,15 +1,18 @@
 # TODO list
 
+## VDU module
+- Use the full font.
+- Tidy up. add initialise.
+- Test the various functions.
+- Add screen read.
+
+## Update Bully
+
 ## ScreenEditor module
 
 - Remove all reads *except* the final extraction.
 - Double height support.
 - Set ink, height options etc.
-
-## VDU module
-
-
-
 ## Computer module
 - The base computer system
