+++ documents/modules.md
@@ -10,17 +10,16 @@ Modules outside the lowest level are abstracted from the Pico API. It is perfect
 
 These are the current modules. In the runtime the functionality of the low level modules is partly handled by the runtime itself, to provide a functionally equivalent interface at a mid-level.
 
-|    Module    | Low  | Purpose                                                      |      |
-| :----------: | ---- | ------------------------------------------------------------ | ---- |
-|    Common    | Yes  | Provides some common hardware functions and logging facilities. |      |
-|     DVI      | Yes  | Lowest level possible DVI interface provides                 |      |
-|     USB      | Yes  | Provides low level HID device interface (e.g. USB packets) and a simple File system. |      |
-|    PSRAM     | Yes  | This is the physical hardware interface to the PSRAM Chip    |      |
-|    Input     |      | HID Manager. Converts the Keyboard HID data to a keyboard queue/tracking system in ASCII with localisation. The Gamepad HID is converted to an easy interface, with a keyboard option if no Gamepad is available. The mouse HID is converted into position and button tracking |      |
-| Alt_Graphics |      | Low level graphics functions - draws rectangles, ellipses, text, lines and similar. This has now been superceded and will not be developed further, but is still available. |      |
-|    Screen    |      | Provides a text i/o layer and full screen editor for text input rather like the Commodore machines. |      |
-|    Bully     |      | This is an application which allows me to 'bully' the USB system (and other things) to see if they crash. |      |
-|    Memory    |      | This manages static RAM in the Pico and PSRAM.               |      |
+| Module | Low  | Purpose                                                      |      |
+| :----: | ---- | ------------------------------------------------------------ | ---- |
+| Common | Yes  | Provides some common hardware functions and logging facilities. |      |
+|  DVI   | Yes  | Lowest level possible DVI interface provides                 |      |
+|  USB   | Yes  | Provides low level HID device interface (e.g. USB packets) and a simple File system. |      |
+| PSRAM  | Yes  | This is the physical hardware interface to the PSRAM Chip    |      |
+| Input  |      | HID Manager. Converts the Keyboard HID data to a keyboard queue/tracking system in ASCII with localisation. The Gamepad HID is converted to an easy interface, with a keyboard option if no Gamepad is available. The mouse HID is converted into position and button tracking |      |
+| Screen |      | Provides a text i/o layer and full screen editor for text input rather like the Commodore machines. |      |
+| Bully  |      | This is an application which allows me to 'bully' the USB system (and other things) to see if they crash. |      |
+| Memory |      | This manages static RAM in the Pico and PSRAM.               |      |
 
 ## Elements of Modules
 
@@ -102,4 +101,4 @@ This may need to be changed so that the environment directory can be found, if i
 
 Paul Robson 
 
-10 August 2025
+13 August 2025
