+++ other/old/alt_graphics/app/main.c
