+++ other/old/alt_graphics/dependencies.info
