+++ other/old/alt_graphics/include/alt_graphics_module_local.h
