+++ other/old/alt_graphics/library/dispatch.c
