+++ other/old/alt_graphics/library/drawing/character.c
