+++ other/old/alt_graphics/library/drawing/ellipse.c
