+++ other/old/alt_graphics/library/drawing/line.c
