+++ other/old/alt_graphics/library/drawing/rect.c
