+++ other/old/alt_graphics/library/drawing/triangle.c
