+++ other/old/alt_graphics/library/raw.c
