+++ other/old/alt_graphics/library/setup.c
