+++ other/old/alt_graphics/library/systemfont.c
